Refactor #9 - exception 패키지에서 commons 패키지로 이동
