+++ src/main/java/me/honki12345/wantedassignment/common/ErrorCode.java
@@ -1,4 +1,4 @@
-package me.honki12345.wantedassignment.exception;
+package me.honki12345.wantedassignment.common;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
